Alteração genérica, inclusão da tabela compras

+++ Loja1.0/Model/Vendas.cs
@@ -29,7 +29,7 @@ namespace Loja1._0.Model
         public Nullable<double> valor_Venda { get; set; }
         public int id_Usuario { get; set; }
         public Nullable<int> id_Cliente { get; set; }
-        public int desconto { get; set; }
+        public decimal desconto { get; set; }
         public Nullable<decimal> comissao { get; set; }
     
         public virtual Clientes Clientes { get; set; }
